Move training and evaluation functions to a separate file
Also adds support for obtaining a coco API object from our dataset
