+++ references/detection/utils.py
@@ -222,6 +222,17 @@ def collate_fn(batch):
     return tuple(zip(*batch))
 
 
+def warmup_lr_scheduler(optimizer, warmup_iters, warmup_factor):
+
+    def f(x):
+        if x >= warmup_iters:
+            return 1
+        alpha = float(x) / warmup_iters
+        return warmup_factor * (1 - alpha) + alpha
+
+    return torch.optim.lr_scheduler.LambdaLR(optimizer, f)
+
+
 def mkdir(path):
     try:
         os.makedirs(path)
